keep moving contracts to core folder
