+++ app/api/common.v2/database/MongoTransactionManager.ts
@@ -1,6 +1,6 @@
 import { MongoClient, ClientSession } from 'mongodb';
 import { Logger } from 'api/log.v2/contracts/Logger';
-import { TransactionManager } from '../contracts/TransactionManager';
+import { TransactionManager } from '../../core/libs/TransactionManager';
 
 export class MongoTransactionManager implements TransactionManager {
   private mongoClient: MongoClient;
