+++ app/api/paragraphExtraction/application/PXCreateExtractor.ts
@@ -1,7 +1,7 @@
 import { TemplatesDataSource } from 'api/core/domain/template/TemplatesDataSource';
-import { IdGenerator } from 'api/common.v2/contracts/IdGenerator';
+import { IdGenerator } from 'api/core/libs/IdGenerator';
 import relationshipTypeDS from 'api/relationtypes';
-import { TransactionManager } from 'api/common.v2/contracts/TransactionManager';
+import { TransactionManager } from 'api/core/libs/TransactionManager';
 import { UseCase } from 'api/core/libs/UseCase';
 import { JobsDispatcher } from 'api/core/libs/queue/application/contracts/JobsDispatcher';
 import { CreateParagraphExtractionEntityStatusesJob } from '../jobs/CreateParagraphExtractionEntityStatusesJob';
