+++ app/api/paragraphExtraction/application/PXDeleteExtractor.ts
@@ -1,5 +1,5 @@
 import { UseCase } from 'api/core/libs/UseCase';
-import { TransactionManager } from 'api/common.v2/contracts/TransactionManager';
+import { TransactionManager } from 'api/core/libs/TransactionManager';
 import { PXExtractorsDataSource } from '../domain/PXExtractorDataSource';
 
 type Input = {
