+++ app/api/relationships.v2/services/DeleteRelationshipService.ts
@@ -1,5 +1,5 @@
 import { AuthorizationService } from 'api/authorization.v2/services/AuthorizationService';
-import { TransactionManager } from 'api/common.v2/contracts/TransactionManager';
+import { TransactionManager } from 'api/core/libs/TransactionManager';
 import { RelationshipsDataSource } from '../contracts/RelationshipsDataSource';
 import { MissingRelationshipError } from '../errors/relationshipErrors';
 import { DenormalizationService } from './DenormalizationService';
