+++ app/api/relationships.v2/services/propertyUpdateStrategies/specs/UpdateRelationshipPropertiesJob.spec.ts
@@ -1,6 +1,6 @@
 import { partialImplementation } from 'api/common.v2/testing/partialImplementation';
 import { EntityRelationshipsUpdateService } from 'api/entities.v2/services/EntityRelationshipsUpdateService';
-import { TransactionManager } from 'api/common.v2/contracts/TransactionManager';
+import { TransactionManager } from 'api/core/libs/TransactionManager';
 import { UpdateRelationshipPropertiesJob } from '../UpdateRelationshipPropertiesJob';
 
 it('should execute the updater and reindex the entity', async () => {
