+++ app/api/templates.v2/services/CreateTemplateService.ts
@@ -6,7 +6,7 @@ import { RelationshipTypesDataSource } from 'api/relationshiptypes.v2/contracts/
 import { RelationshipPropertyData } from 'shared/types/api.v2/templates.createTemplateRequest';
 import { createError } from 'api/utils';
 import { DenormalizationService } from 'api/relationships.v2/services/DenormalizationService';
-import { TransactionManager } from 'api/common.v2/contracts/TransactionManager';
+import { TransactionManager } from 'api/core/libs/TransactionManager';
 import { TemplatesDataSource } from '../../core/domain/template/TemplatesDataSource';
 import { QueryMapper } from '../../core/infrastructure/mongodb/template/QueryMapper';
 import { TemplateInput, TemplateInputMappers } from './TemplateInputMappers';
